+++ bizsocket-base/src/main/java/bizsocket/base/Body.java
@@ -1,14 +1,15 @@
-package bizsocket.rx;
+package bizsocket.base;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
+
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Documented
 @Target(PARAMETER)
 @Retention(RUNTIME)
-public @interface QueryMap {
+public @interface Body {
 
 }
