+++ bizsocket-base/src/main/java/bizsocket/base/JSONResponseConverter.java
@@ -1,4 +1,4 @@
-package bizsocket.rx;
+package bizsocket.base;
 
 import bizsocket.tcp.Packet;
 import okio.ByteString;
