+++ bizsocket-base/src/main/java/bizsocket/base/Query.java
@@ -1,4 +1,4 @@
-package bizsocket.rx;
+package bizsocket.base;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
