+++ bizsocket-base/src/main/java/bizsocket/base/QueryMap.java
@@ -1,15 +1,14 @@
-package bizsocket.rx;
+package bizsocket.base;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
-
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Documented
 @Target(PARAMETER)
 @Retention(RUNTIME)
-public @interface Body {
+public @interface QueryMap {
 
 }
