+++ bizsocket-base/src/test/java/bizsocket/rx/BizSocketRxSupportTest.java
@@ -1,5 +1,8 @@
 package bizsocket.rx;
 
+import bizsocket.base.BizSocketRxSupport;
+import bizsocket.base.JSONRequestConverter;
+import bizsocket.base.JSONResponseConverter;
 import bizsocket.core.Configuration;
 import client.SampleSocketClient;
 import junit.framework.TestCase;
