+++ bizsocket-base/src/test/java/bizsocket/rx/JSONRequestConverterTest.java
@@ -1,6 +1,12 @@
 package bizsocket.rx;
 
 import junit.framework.TestCase;
+
+import bizsocket.base.JSONRequestConverter;
+import bizsocket.base.Query;
+import bizsocket.base.QueryMap;
+import bizsocket.base.Request;
+import bizsocket.base.Tag;
 import okio.ByteString;
 import org.json.JSONObject;
 import org.junit.Test;
