+++ bizsocket-base/src/test/java/client/SamplePacketFactory.java
