+++ bizsocket-base/src/test/java/common/SampleBizPacketValidator.java
