+++ bizsocket-base/src/test/java/common/SampleCmd.java
