+++ bizsocket-base/src/test/java/common/SamplePacket.java
