+++ bizsocket-base/src/test/java/common/SampleProtocolUtil.java
