+++ sample/j2se/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    compile project(':bizsocket-rx')
+    compile project(':bizsocket-base')
 
     compile 'org.json:org.json:2.0'
     testCompile group: 'junit', name: 'junit', version: '4.11'
